+++ build.gradle
@@ -65,7 +65,7 @@ dependencies {
 
     implementation 'org.bouncycastle:bcprov-jdk18on:1.75'
 
-    // Kubernetes
+    // Kubernetes java client
     implementation 'io.kubernetes:client-java:20.0.1'
 }
