+++ src/main/resources/application.yml
@@ -7,8 +7,8 @@ spring:
 
 ---
 
-spring:
-  profiles:
-    active: dev
+#spring:
+#  profiles:
+#    active: dev
 
 ---
